requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
blinker==1.9.0
    # via flask
certifi==2026.2.25
    # via requests
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via flask
flask==3.1.3
    # via -r requirements.in
gunicorn==25.2.0
    # via -r requirements.in
idna==3.11
    # via requests
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
packaging==26.0
    # via gunicorn
psycopg2-binary==2.9.11
    # via -r requirements.in
requests==2.33.0
    # via -r requirements.in
urllib3==2.6.3
    # via requests
werkzeug==3.1.6
    # via flask
